requirements.txt
pyTelegramBotAPI==4.15.0
python-dotenv==1.0.1
